# o_donjon/public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API O'donjon",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://odonjon.fr/",
            "description": "API pour O'donjon"
        }
    ],
    "paths": {
...
        "/campaign/{id}/npc": {
            "get": {
                "tags": [
                    "Campaign"
                ],
                "operationId": "App\\Controller\\CampaignController::browseNpc",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of NPCs of a campaign",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/campaignNPCs"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
...
        "/nps": {
            "get": {
                "tags": [
                    "NPC"
                ],
                "operationId": "App\\Controller\\NpcController::browse",
                "responses": {
                    "200": {
                        "description": "List of NPCs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/npcInfo"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/npc": {
            "post": {
                "tags": [
                    "NPC"
                ],
                "operationId": "App\\Controller\\NpcController::add",
                "requestBody": {
                    "$ref": "#/components/requestBodies/npc"
                },
                "responses": {
                    "200": {
                        "description": "Informations of the NPC",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/npcInfo"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/npc/{id}": {
            "get": {
                "tags": [
                    "NPC"
                ],
                "operationId": "App\\Controller\\NpcController::read",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informations of the NPC",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/npcInfo"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            },
            "put": {
                "tags": [
                    "NPC"
                ],
                "operationId": "App\\Controller\\NpcController::edit",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/npc"
                },
                "responses": {
                    "200": {
                        "description": "Informations of the NPC",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/npcInfo"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "NPC"
                ],
                "operationId": "App\\Controller\\NpcController::delete",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "campaignNPCs": {
                "description": "List of NPCs of the campaign",
                "properties": {
                    "NPCs": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/npc"
                        }
                    }
                },
                "type": "object"
            },
...
            "npc": {
                "description": "Information of a NPC",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "7"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 64,
                        "example": "Cocotte",
                        "nullable": true
                    },
                    "type": {
                        "type": "string",
                        "maxLength": 64,
                        "example": "animal",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Au mon dieu une poule qui parle !",
                        "nullable": true
                    },
                    "isAlly": {
                        "type": "string",
                        "maxLength": 64,
                        "example": "true",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "npcInfo": {
                "description": "Information of a NPC",
                "properties": {
                    "campaign": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": "7"
                            },
                            "name": {
                                "type": "string",
                                "maxLength": 64,
                                "example": "Les poules de ma grand-mère",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/npc"
                    }
                ]
            },
...
        },
        "responses": {
            "notFound": {
                "description": "ressource doesn't existe"
            }
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "description": "ID of the ressource",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            }
        },
        "requestBodies": {
...
            "npc": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "name",
                                "type",
                                "description",
                                "isAlly",
                                "campaign"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "maxLength": 64,
                                    "example": "Cocotte",
                                    "nullable": true
                                },
                                "type": {
                                    "type": "string",
                                    "maxLength": 64,
                                    "example": "animal",
                                    "nullable": true
                                },
                                "description": {
                                    "type": "string",
                                    "example": "Au mon dieu une poule qui parle !",
                                    "nullable": true
                                },
                                "isAlly": {
                                    "type": "string",
                                    "maxLength": 64,
                                    "example": "true",
                                    "nullable": true
                                },
                                "campaign": {
                                    "type": "integer",
                                    "example": "7"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "bearerFormat": "JWT"
            }
        }
    }
}